cli: add enable_logging verb
By default, when generating security files, a logging configuration is
not generated. This is because enabling the security log actually
removes security events from the standard log, which only makes sense if
one is intending to consume them in another way (e.g. a SIEM).

Add a new CLI verb `enable_logging` to optionally enable logging for a
given identity, or for all identities in the keystore.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>